Performance tests: Log timings of performance test segments
The actual running of the performance tests accounts for almost all of the time
that the GitHub Actions step takes for the performance test check. In this patch
we're adding log points to report the individual runtimes for each step in that
testing process, giving more insight into where our time is spent and where we
might be able to address the poor runtime performance.